Fix for hue and small hashes
Hue of 0% is the same color as hue of 100%. Anyways hue is usually represented as an angle [0° to 360°) because it is cyclical or in this case a percentage of a circle [0%, 100%).

Using the first 7 characters and last 4 characters of the hash because if a short hash is given then the last 7 characters run into the first 15 characters. Which causes the color and pixels to be related. Having hue first limits the effect of the overlap with really small hashes.

There should probably be a warning if the hash is not hex or is less than 11 characters (bare minimum is 7 characters).